17 & 24: rework PowerSet iter adaptor
I finally figured out what was going on with closures and trying to set
a default value when using a basic constructor vs providing a closure/fn
from a different constructor.

Structure update syntax requires updating from the same type (makes
sense), but every closure is it's own type so when monomorphizing a Fn
type parameter you can't use update syntax to replace the default
closure with a different one.  The only way to use update syntax for
this is to use trait objects and dynamic dispatch.  Which is slightly
slower, but structures the code the way I want it.