day24: fix bug from over aggressive optimization
Trying to cut down the search space by QE (product) was a mistake, it
can't ignore a solution with a smaller QE if the length is smaller
(leg-room first!). Getting it right wouldn't help, at most it would
drop a few solutions early but not cut down the search at all.

Adding in the test cases caught this :)