more bikeshedding on 24
Don't collect all the possible solutions, we don't need them all at once. It is
possible to search on multiple attributes (len & product), as long as they are
placed into a tuple in priority order min works just fine.

min_by_key would be appropriate if we needed to keep the vector, but the answer
we want is one of the calculated components of our key so map->min->destructure